ログインしているか確認するタスクコントローラの処理を共通部品に移行
