+++ task_management/app/controllers/application_controller.rb
@@ -2,6 +2,10 @@
 
 # アプリケーションコントローラー
 class ApplicationController < ActionController::Base
+  include Common
+
+  helper_method :current_user
+
   before_action :render_maintenance, if: :maintenance_mode?
 
   unless Rails.env.development?
@@ -38,25 +42,5 @@ class ApplicationController < ActionController::Base
   def render_500(exception = nil)
     logger.info "Rendering 500 with exception: #{exception.message}" if exception
     render 'errors/500', status: :internal_server_error
-  end
-
-  private
-
-  def log_in(user)
-    session[:user_id] = user.id
-  end
-
-  def log_out
-    session.delete(:user_id) if session[:user_id]
-  end
-
-  helper_method :current_user
-
-  def current_user
-    @current_user ||= User.find_by(id: session[:user_id])
-  end
-
-  def logged_in?
-    current_user.present?
   end
 end
